refactor: extract editor UI state into Zustand store
Move scattered useState calls in EditorPanel into a dedicated
editorUIStore (Zustand), add colorUtils helper, and wire Toolbar
to consume the same store. Adds zustand ^5.0.12 as a dependency.
